break out tests for nodes and helpers into separate files
